Refactor logic out of query builder
